requirements.txt
# requirements.txt - Required packages for MemoryPal Enhanced V2

# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0
numpy>=1.24.0
pathlib
datetime
typing
json
re
hashlib
warnings

# Document processing
PyMuPDF>=1.23.0  # for PDF processing (fitz)
beautifulsoup4>=4.12.0

# Google Gemini API
google-generativeai>=0.3.0

# Optional: Supabase (for database functionality)
supabase>=2.0.0

# Audio processing dependencies
audio-recorder-streamlit>=0.0.8
whisper>=1.1.10  # OpenAI Whisper for transcription
librosa>=0.10.0  # Audio processing
soundfile>=0.12.1  # Audio file I/O
speech-recognition>=3.10.0  # Alternative speech recognition

# Text-to-Speech
pyttsx3>=2.90  # For podcast generation

# Advanced NLP and ML
transformers>=4.30.0
torch>=2.0.0  # For transformers
pyannote.audio>=3.1.0  # For speaker diarization

# Optional: Cohere for reranking (if you have API key)
cohere>=4.30.0

# Optional: For advanced audio processing
scipy>=1.10.0
matplotlib>=3.7.0  # For visualizations

# Utility packages
requests>=2.31.0
urllib3>=2.0.0
pathlib2  # For older Python versions
tqdm>=4.65.0  # Progress bars

# Optional: For better keyword extraction
scikit-learn>=1.3.0
nltk>=3.8.0

# Optional: For improved PDF processing
pymupdf4llm>=0.0.5

# Development and testing (optional)
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0

# Note: Some packages might require additional system dependencies:
# - ffmpeg (for audio processing): 
#   Ubuntu/Debian: sudo apt update && sudo apt install ffmpeg
#   macOS: brew install ffmpeg
#   Windows: Download from https://ffmpeg.org/
#
# - For pyannote.audio, you'll need a Hugging Face token:
#   Set HUGGINGFACE_TOKEN in your .env file
#   Get token from: https://huggingface.co/settings/tokens